fix: resolve remaining int/double mismatches in POS page

+++ lib/features/billing/data/kiosk_service.dart
@@ -15,7 +15,7 @@ class KioskProductResult {
   final double pricePerKg;
   final bool isUnit;
   final String? unitName;
-  final int unitMultiplier;
+  final double unitMultiplier;
   final double singlePrice;
   final double savings;
   final String? imageUrl;
